pyproject.toml
[project]
name = "openai-agents-testkit"
version = "0.1.0"
description = "Testing utilities for openai-agents-python: fake models, providers, and pytest fixtures"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "xncbf", email = "xncbf12@gmail.com" }]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    "Framework :: Pytest",
    "Framework :: AsyncIO",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Testing :: Mocking",
]
keywords = ["openai", "agents", "testing", "mock", "fake", "pytest", "fixtures"]

dependencies = [
    "openai-agents>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "ty>=0.0.1a7",
    "pre-commit>=4.0.0",
]
docs = [
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.27.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/xncbf/openai-agents-testkit"
Documentation = "https://xncbf.github.io/openai-agents-testkit/"
Repository = "https://github.com/xncbf/openai-agents-testkit"
Issues = "https://github.com/xncbf/openai-agents-testkit/issues"
Changelog = "https://github.com/xncbf/openai-agents-testkit/blob/main/CHANGELOG.md"

[project.entry-points.pytest11]
openai_agents_testkit = "openai_agents_testkit.fixtures"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openai_agents_testkit"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/README.md",
    "/LICENSE",
]

[dependency-groups]
dev = [
    "openai-agents-testkit[dev]",
]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "PTH",  # flake8-use-pathlib
    "RUF",  # Ruff-specific rules
    "ASYNC",# flake8-async
]
ignore = [
    "E501", # line too long (handled by formatter)
    "B008", # do not perform function calls in argument defaults
    "B904", # allow raise without from in except
]

[tool.ruff.lint.isort]
known-first-party = ["openai_agents_testkit"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ARG"]
"src/openai_agents_testkit/models.py" = ["ARG"]  # Interface implementation requires unused args

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]

[tool.coverage.run]
source = ["src/openai_agents_testkit"]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
    "@overload",
]
show_missing = true

[tool.ty.environment]
python-version = "3.10"
